Refactor config, add modal extension
